Add more options to configure italicization
Before this commit, it was possible to enable italicization of strings
and comments. This seems fairly arbitrary as users (myself included) may
want to choose settings other than these. For instance, some other
themes use italicization to separate keywords, or other
language-specific patterns. While this may not be of interest to
everyone, it seems reasonable to allow it to be configurable.

This patch therefore introduces the following configuration options:
- g:gruvbox_italicize_statements
- g:gruvbox_italicize_conditionals
- g:gruvbox_italicize_repeats
- g:gruvbox_italicize_label
- g:gruvbox_italicize_exceptions
- g:gruvbox_italicize_keywords